chore: extract contentful html parsing to component ContentfulHtml
